requirements.txt
annotated-types==0.7.0
anyio==3.7.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.4
docx==0.2.4
fastapi==0.104.1
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.188.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httplib2==0.31.2
idna==3.11
lxml==6.0.2
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.1.0
proto-plus==1.27.0
protobuf==5.29.5
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.5.0
pydantic_core==2.14.1
pyparsing==3.3.2
pypdfium2==5.3.0
python-docx==1.2.0
python-dotenv==1.2.1
python-multipart==0.0.22
requests==2.32.5
rsa==4.9.1
sniffio==1.3.1
starlette==0.27.0
tqdm==4.67.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.24.0
